Errorhandle (#3)
* bug fix for forstmt and ifstmt

* disable info

* add test points for error handle

* add position info in ast

* add error code

* use position instead of offset

* add checker and fix code style

* add check class

* cache

* CACHE

* complete array init check

* cache

* able to check expr

* check ok

* impl error handle

* add checker

+++ scanner/scanner.cpp
@@ -169,7 +169,7 @@ void Scanner::Scan(int *pos, token::Token *tok, string *lit) {
 }
 
 void Scanner::Next() {
-    if (read_offset_ >= src_.size()) {
+    if (read_offset_ >= (int)src_.size()) {
         offset_ = (int)src_.size();
         if (ch_ == '\n') {
             file_->AddLine(offset_);
